Extract one more String method
