+++ src/test/groovy/edu/ie3/datamodel/io/FileNamingStrategyTest.groovy
@@ -61,75 +61,6 @@ import spock.lang.Specification
 
 class FileNamingStrategyTest extends Specification {
 
-	def "The FileNamingStrategy is capable of cleaning up strings correctly"() {
-		when:
-		String actual = FileNamingStrategy.cleanString(input)
-
-		then:
-		actual == expected
-
-		where:
-		input 		|| expected
-		"ab123" 	|| "ab123"
-		"ab.123" 	|| "ab_123"
-		"ab-123" 	|| "ab_123"
-		"ab_123" 	|| "ab_123"
-		"ab/123" 	|| "ab_123"
-		"ab\\123" 	|| "ab_123"
-		"ab!123" 	|| "ab_123"
-		"ab\"123" 	|| "ab_123"
-		"ab§123" 	|| "ab_123"
-		"ab\$123" 	|| "ab_123"
-		"ab&123" 	|| "ab_123"
-		"ab{123" 	|| "ab_123"
-		"ab[123" 	|| "ab_123"
-		"ab}123" 	|| "ab_123"
-		"ab]123" 	|| "ab_123"
-		"ab(123" 	|| "ab_123"
-		"ab)123" 	|| "ab_123"
-		"ab=123" 	|| "ab_123"
-		"ab?123" 	|| "ab_123"
-		"abß123" 	|| "ab_123"
-		"ab123." 	|| "ab123_"
-		"ab123-" 	|| "ab123_"
-		"ab123_" 	|| "ab123_"
-		"ab123/" 	|| "ab123_"
-		"ab123\\" 	|| "ab123_"
-		"ab123!" 	|| "ab123_"
-		"ab123\"" 	|| "ab123_"
-		"ab123§" 	|| "ab123_"
-		"ab123\$" 	|| "ab123_"
-		"ab123&" 	|| "ab123_"
-		"ab123{" 	|| "ab123_"
-		"ab123[" 	|| "ab123_"
-		"ab123}" 	|| "ab123_"
-		"ab123]" 	|| "ab123_"
-		"ab123(" 	|| "ab123_"
-		"ab123)" 	|| "ab123_"
-		"ab123=" 	|| "ab123_"
-		"ab123?" 	|| "ab123_"
-		"ab123ß" 	|| "ab123_"
-		".ab123" 	|| "_ab123"
-		"-ab123" 	|| "_ab123"
-		"_ab123" 	|| "_ab123"
-		"/ab123" 	|| "_ab123"
-		"\\ab123" 	|| "_ab123"
-		"!ab123" 	|| "_ab123"
-		"\"ab123" 	|| "_ab123"
-		"§ab123" 	|| "_ab123"
-		"\$ab123"	|| "_ab123"
-		"&ab123" 	|| "_ab123"
-		"{ab123" 	|| "_ab123"
-		"[ab123" 	|| "_ab123"
-		"}ab123" 	|| "_ab123"
-		"]ab123" 	|| "_ab123"
-		"(ab123" 	|| "_ab123"
-		")ab123" 	|| "_ab123"
-		"=ab123" 	|| "_ab123"
-		"?ab123" 	|| "_ab123"
-		"ßab123" 	|| "_ab123"
-	}
-
 	def "The FileNamingStrategy is able to prepare the prefix properly"() {
 		when:
 		String actual = FileNamingStrategy.preparePrefix(prefix)
